% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_assign.R
\name{alignAssignArrow}
\alias{alignAssignArrow}
\title{Align a highlighted region's assignment operators.}
\usage{
alignAssignArrow()
}
\value{
Aligns the single caret operators (\code{<-}) within a
highlighted region.
}
\description{
Align a highlighted region's assignment operators.
}
