% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_assign.R
\name{alignAssignComment}
\alias{alignAssignComment}
\title{Align a highlighted region's comments.}
\usage{
alignAssignComment()
}
\value{
Aligns the comments (\code{#}) within a
highlighted region.
}
\description{
Align a highlighted region's comments.
}
