% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_assign.R
\name{alignAssignEqual}
\alias{alignAssignEqual}
\title{Align a highlighted region's assignment operators.}
\usage{
alignAssignEqual()
}
\value{
Aligns the equal sign assignment operators (\code{=}) within a
highlighted region.
}
\description{
Align a highlighted region's assignment operators.
}
