% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_assign.R
\name{alignAssignHashtag}
\alias{alignAssignHashtag}
\title{#' Align a highlighted region's assignment operators.
#'
#' @return Aligns the equal sign assignment operators (\code{=}) within a
#' highlighted region.
#' @export
alignAssignEqual <- function() {
  alignAssign("=")
}}
\usage{
alignAssignHashtag()
}
\value{
Aligns the hastags (\code{#}) within a
highlighted region.
}
\description{
#' Align a highlighted region's assignment operators.
#'
#' @return Aligns the single caret operators (\code{<-}) within a
#' highlighted region.
#' @export
alignAssignArrow <- function() {
  alignAssign("<-")
}
Align a highlighted region's comment hastags.
}
