% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_assign.R
\name{alignAssignStanComment}
\alias{alignAssignStanComment}
\title{Align a highlighted region's Stan-formatted comments.}
\usage{
alignAssignStanComment()
}
\value{
Aligns the comments (\code{#}) within a
highlighted region.
}
\description{
Align a highlighted region's comments.
}
