% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_assign.R
\name{alignr}
\alias{alignr}
\title{Align a highlighted region's assignment operators.}
\usage{
alignr()
}
\value{
Aligns the single caret operators (\code{<-}) with a highlighted region.
}
\description{
Align a highlighted region's assignment operators.
}

